requirements.txt
asgiref==3.2.10
Django==3.1
django-smart-selects==1.5.9
gunicorn==20.0.4
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.2.0
